assignment-1/cloud/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
DateTime==4.3
google-cloud==0.34.0
firebase-admin==4.5.3